+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial support for embeddings/vectors
 - Support for dimensionality reduction
 - Initial implementation of MMD-Critic
+- Initial implementation of labelwise MMD-Critic
 
 ## [0.3.8] - 2021-09-07
 ### Changed